refactor: isolate home screen
